Feature#TestCode - 읽기 작업 동기 쓰기 작업 비동기 처리 - 한 작업 내에서 읽기 작업과 쓰기 작업을 분리 - Service.class ->Service.Interface
* 유닛 테스트 코드
- unit Package - 모두 통과
- CpuMonitoringManageService 83%
- CpuMonitoringBackupService 100%
- CpuMonitoringService 100%

* 통합 테스트 코드
- total Package - 모두 통과
- CpuMonitoringManageService 88%
- CpuMonitoringBackupService 64%
- CpuMonitoringService 88%

+++ src/main/java/com/tera/pretest/context/cpumonitoring/repository/base/CpuUsageRateByHourRepository.java
@@ -16,7 +16,7 @@ public interface CpuUsageRateByHourRepository extends JpaRepository<CpuUsageRate
 
     @Modifying
     @Query("UPDATE CpuUsageRateByHour stats SET stats.flag = 'Y' WHERE stats.createTime < :pastDay")
-    long softDeleteOldData(@Param("pastDay") ZonedDateTime pastDay);
+    int softDeleteOldData(@Param("pastDay") ZonedDateTime pastDay);
 
     List<CpuUsageRateByHour> findByFlag(String flag);
